+++ pkg/service/audit_utils.go
@@ -112,6 +112,25 @@ func CreateUserAuditEvent(ctx context.Context, al *zap.Logger, db bun.IDB, actio
 	}
 }
 
+func CreateUserLoginAuditEvent(ctx context.Context, al *zap.Logger, action string, name string) {
+	sd, ok := GetSessionDataFromContext(ctx)
+	if !ok {
+		_log.Warn("unable to create audit event: could not fetch info from context")
+		return
+	}
+
+	detail := &audit.EventDetail{
+		Message: fmt.Sprintf("User login: %s", name),
+		Meta: map[string]string{
+			"user": name,
+		},
+	}
+	if err := audit.CreateV1Event(al, sd, detail, fmt.Sprintf("user.%s.success", action), ""); err != nil {
+		_log.Warn("unable to create audit event", err)
+	}
+
+}
+
 func CreateGroupAuditEvent(ctx context.Context, al *zap.Logger, db bun.IDB, action string, name string, id uuid.UUID, usersBefore, usersAfter, rolesBefore, rolesAfter []uuid.UUID) {
 	sd, ok := GetSessionDataFromContext(ctx)
 	if !ok {